added mixamo zombiegirl as non-uma zombie example
