+++ Assets/Models/mixamo.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f601633ed3de3964b9681e56590c9663
+folderAsset: yes
+timeCreated: 1522846772
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
