+++ Assets/Models/mixamo/Materials.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 7185af962851e074d9332cc04df6890b
+folderAsset: yes
+timeCreated: 1522846772
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
