+++ Assets/Models/mixamo/Materials/Jill_HiMidRes_diffuse.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b89cf8865e43774bbccc7e08326822a
+timeCreated: 1437288268
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
