+++ Assets/Models/mixamo/Materials/Milton_midres_diffuse.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be7307cdd09e7f14ea2d37b2cacd45fb
+timeCreated: 1437288249
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
