+++ Assets/Models/mixamo/Materials/Police Zombie_diffuse.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0ae8ba98626a9ec42870c213a7549d68
+timeCreated: 1437288261
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
