+++ Assets/Models/mixamo/Materials/Zoe_Clothes_COL.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a5528bbca00c92542b06313966d86faa
+timeCreated: 1437288278
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
