+++ Assets/Models/mixamo/Materials/Zoe_Skin_COL.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ba83606706b70eb4c94bee8d43ac91b7
+timeCreated: 1437288273
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
