+++ Assets/Models/mixamo/Materials/face.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 29a3a338b35618c459e22af5989bcef1
+timeCreated: 1437235577
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
