+++ Assets/Models/mixamo/Materials/glasses.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3a7a3e9e9a9c5ab4296ccb8b2fc6f0a5
+timeCreated: 1437235577
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
