+++ Assets/Models/mixamo/Materials/shoes.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 24dad75dd2f306e4daaa79ce83313053
+timeCreated: 1437235577
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
