+++ Assets/Models/mixamo/Materials/zombie_diffuse.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d9beb60691d225048bc0f6a43e4e5a2c
+timeCreated: 1437288241
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
