+++ Assets/Models/mixamo/Zombiegirl_W_Kurniawan@zombie_walk.fbm.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 56629d608de833349a0e662e5ad192f9
+folderAsset: yes
+timeCreated: 1522846772
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
