+++ Assets/Prefabs/zombies.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: e627c378600f3aa4787547cd2f62542a
+folderAsset: yes
+timeCreated: 1522848947
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
