+++ Assets/Prefabs/zombies/non_uma_zombie.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 6f86cc4ca860026448486021037c9e47
+timeCreated: 1522848955
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
