--- Assets/Scripts/ragdoll/RagdollCreatorTest.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 38f6de6550aa7714ab8197ee9c3b2569
-timeCreated: 1466447622
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
